# BackendServices/matrimonialwebapi/src/MatrimonialApi/Models/CommunicationList.cs
/*
 * Matrimonial API - OpenAPI 3.0
 *
 * Design and definition of Matrimonial APIs created for practice and teaching
 *
 * OpenAPI spec version: 1.0.11
 * Contact: floatingrays@gmail.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class CommunicationList : IEquatable<CommunicationList>
    { 
        /// <summary>
        /// Gets or Sets CurrentPageNumber
        /// </summary>

        [DataMember(Name="currentPageNumber")]
        public int? CurrentPageNumber { get; set; }

        /// <summary>
        /// Gets or Sets Range
        /// </summary>

        [DataMember(Name="range")]
        public int? Range { get; set; }

        /// <summary>
        /// Gets or Sets Communications
        /// </summary>

        [DataMember(Name="Communications")]
        public List<Communication> Communications { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class CommunicationList {\n");
            sb.Append("  CurrentPageNumber: ").Append(CurrentPageNumber).Append("\n");
            sb.Append("  Range: ").Append(Range).Append("\n");
            sb.Append("  Communications: ").Append(Communications).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((CommunicationList)obj);
        }

        /// <summary>
        /// Returns true if CommunicationList instances are equal
        /// </summary>
        /// <param name="other">Instance of CommunicationList to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(CommunicationList other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    CurrentPageNumber == other.CurrentPageNumber ||
                    CurrentPageNumber != null &&
                    CurrentPageNumber.Equals(other.CurrentPageNumber)
                ) && 
                (
                    Range == other.Range ||
                    Range != null &&
                    Range.Equals(other.Range)
                ) && 
                (
                    Communications == other.Communications ||
                    Communications != null &&
                    Communications.SequenceEqual(other.Communications)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (CurrentPageNumber != null)
                    hashCode = hashCode * 59 + CurrentPageNumber.GetHashCode();
                    if (Range != null)
                    hashCode = hashCode * 59 + Range.GetHashCode();
                    if (Communications != null)
                    hashCode = hashCode * 59 + Communications.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(CommunicationList left, CommunicationList right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(CommunicationList left, CommunicationList right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}