# BackendServices/matrimonialwebapi/src/MatrimonialApi/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Matrimonial API - OpenAPI 3.0",
    "description" : "Design and definition of Matrimonial APIs created for practice and teaching",
    "contact" : {
      "email" : "floatingrays@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.11"
  },
...
  "servers" : [ {
    "url" : "http://localhost:3002/api"
  } ],
...
  "paths" : {
...
    "/Communication" : {
      "summary" : "Trigger a communication to the preferred profile ",
      "description" : "Trigger a communication to the preferred profile ",
      "get" : {
        "tags" : [ "Communication" ],
        "summary" : "Get all the communication done till date",
        "description" : "Get all the communication done till date",
        "operationId" : "GetCommunication",
        "parameters" : [ {
          "name" : "X-Request-Auth",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "pageNumber",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "range",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CommunicationList"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        }
      },
      "post" : {
        "tags" : [ "Communication" ],
        "summary" : "Add a communication",
        "description" : "Add a communication",
        "operationId" : "AddCommunication",
        "parameters" : [ {
          "name" : "X-Request-Auth",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "Add a communication",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Communication"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CommunicationList"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Communication" : {
        "type" : "object",
        "properties" : {
          "to" : {
            "type" : "string"
          },
          "from" : {
            "type" : "string"
          },
          "commnicationType" : {
            "type" : "string",
            "enum" : [ "email", "sms", "whatsapp" ]
          },
          "message" : {
            "type" : "string"
          },
          "subject" : {
            "type" : "string"
          },
          "createdDateTime" : {
            "type" : "string"
          },
          "messageDeliveredStatus" : {
            "type" : "string",
            "enum" : [ "success", "failed", "processing" ]
          }
        },
        "example" : {
          "commnicationType" : "email",
          "subject" : "subject",
          "createdDateTime" : "createdDateTime",
          "from" : "from",
          "to" : "to",
          "messageDeliveredStatus" : "success",
          "message" : "message"
        }
      },
...
      "CommunicationList" : {
        "type" : "object",
        "properties" : {
          "currentPageNumber" : {
            "type" : "integer"
          },
          "range" : {
            "type" : "integer"
          },
          "Communications" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Communication"
            }
          }
        },
        "example" : {
          "currentPageNumber" : 0,
          "Communications" : [ {
            "commnicationType" : "email",
            "subject" : "subject",
            "createdDateTime" : "createdDateTime",
            "from" : "from",
            "to" : "to",
            "messageDeliveredStatus" : "success",
            "message" : "message"
          }, {
            "commnicationType" : "email",
            "subject" : "subject",
            "createdDateTime" : "createdDateTime",
            "from" : "from",
            "to" : "to",
            "messageDeliveredStatus" : "success",
            "message" : "message"
          } ],
          "range" : 6
        }
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      },
      "BearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "SSO_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://localhost:3002/oauth/authorize",
            "scopes" : {
              "write:pets" : "modify data",
              "read:pets" : "read data"
            }
          }
        }
      },
      "api_key" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  }
}